+++ android/src/main/kotlin/com/yandex/mobile/ads/flutter/util/RequestUtil.kt
@@ -19,7 +19,6 @@ internal fun Map<String, Any?>.toAdRequest() = AdRequest.Builder().also { adRequ
     (getValueOrNull<String>(GENDER))?.let(adRequestBuilder::setGender)
     (getValueOrNull<Map<String, Double>>(LOCATION))?.toLocation()?.let(adRequestBuilder::setLocation)
     (getValueOrNull<Map<String, String>>(PARAMETERS))?.let(adRequestBuilder::setParameters)
-    (getValueOrNull<String>(BIDDING_DATA))?.let(adRequestBuilder::setBiddingData)
     (getValueOrNull<String>(PREFERRED_THEME))?.toAdTheme()?.let(adRequestBuilder::setPreferredTheme)
 }.build()
 
@@ -29,5 +28,4 @@ private const val CONTEXT_TAGS = "contextTags"
 private const val GENDER = "gender"
 private const val LOCATION = "location"
 private const val PARAMETERS = "parameters"
-private const val BIDDING_DATA = "biddingData"
 private const val PREFERRED_THEME = "preferredTheme"