+++ ios/Classes/BannerAd/BannerAdViewFactory.swift
@@ -11,13 +11,15 @@ import Flutter
 import YandexMobileAds
 
 final class BannerAdViewFactory: NSObject, FlutterPlatformViewFactory {
-
+    
     private let messenger: FlutterBinaryMessenger
-
-    init(messenger: FlutterBinaryMessenger) {
+    private let bannerAdSizeFactory: BannerAdSizeFactory
+    
+    init(messenger: FlutterBinaryMessenger, bannerAdSizeFactory: BannerAdSizeFactory) {
         self.messenger = messenger
+        self.bannerAdSizeFactory = bannerAdSizeFactory
     }
-
+    
     func create(
         withFrame frame: CGRect,
         viewIdentifier viewId: Int64,
@@ -26,30 +28,18 @@ final class BannerAdViewFactory: NSObject, FlutterPlatformViewFactory {
         let params = args as? [String: Any?] ?? [:]
         let adUnitId: String = params[.adUnitId] ?? ""
         let id: Int = params[.id] ?? -1
-        let width: CGFloat = params[.width] ?? 0
-        let height: CGFloat = params[.height] ?? 0
-        let type: String = params[.type] ?? ""
-        let adSize: YMABannerAdSize
-
-        switch AdSizeType(rawValue: type) {
-        case .inline:
-            adSize = YMABannerAdSize.inlineSize(withWidth: width, maxHeight: height)
-        case .sticky:
-            adSize = YMABannerAdSize.stickySize(withContainerWidth: width)
-        default:
-            adSize = YMABannerAdSize.inlineSize(withWidth: 0, maxHeight: 0)
-        }
-
+        let adSize = bannerAdSizeFactory.createFromCommandParams(params: params)
+        
         let delegate = BannerAdEventDelegate()
         let name = "\(YandexMobileAdsPlugin.channelName).\(bannerAdChannelName).\(id)"
         let methodChannel = FlutterMethodChannel(name: name, binaryMessenger: messenger)
         let eventChannel = FlutterEventChannel(name: "\(name).events", binaryMessenger: messenger)
         let banner = BannerAdView(
-                adUnitId: adUnitId,
-                adSize: adSize,
-                delegate: delegate,
-                methodChannel: methodChannel,
-                eventChannel: eventChannel
+            adUnitId: adUnitId,
+            adSize: adSize,
+            delegate: delegate,
+            methodChannel: methodChannel,
+            eventChannel: eventChannel
         )
         let provider = BannerAdCommandProvider(banner: banner) {
             methodChannel.setMethodCallHandler(nil)
@@ -57,10 +47,10 @@ final class BannerAdViewFactory: NSObject, FlutterPlatformViewFactory {
         }
         methodChannel.setMethodCallHandler(provider.callHandler)
         eventChannel.setStreamHandler(delegate)
-
+        
         return banner
     }
-
+    
     func createArgsCodec() -> FlutterMessageCodec & NSObjectProtocol {
         FlutterStandardMessageCodec.sharedInstance()
     }
@@ -68,16 +58,12 @@ final class BannerAdViewFactory: NSObject, FlutterPlatformViewFactory {
     private let bannerAdChannelName = "bannerAd"
 }
 
-private enum AdSizeType: String {
-    case inline, sticky
-}
-
-private enum BannerCreationParameter: String {
-    case adUnitId, id, width, height, type
+private enum BannerAdCreationParameter: String {
+    case adUnitId, id
 }
 
 private extension Dictionary where Key == String {
-    subscript<T>(_ key: BannerCreationParameter) -> T? {
+    subscript<T>(_ key: BannerAdCreationParameter) -> T? {
         self[key.rawValue] as? T
     }
 }